requirements.txt
streamlit
pandas 
ydata-profiling
huggingface-hub
python-dotenv